+++ projects/gallery/src/lib/components/viewer/viewer.component.ts
@@ -30,6 +30,7 @@ import {
   OrientationFlag,
   UA,
   VerticalOrientation,
+  ContentTemplateContext,
 } from '../../core';
 import { GalleryItemInternal, isVideo } from '../../core/gallery-item';
 
@@ -64,6 +65,7 @@ export class ViewerComponent implements OnChanges, OnInit, OnDestroy {
   @Input() loadingTemplate: TemplateRef<void>;
   @Input() errorTemplate: TemplateRef<void>;
   @Input() arrowTemplate: TemplateRef<void>;
+  @Input() contentTemplate: TemplateRef<ContentTemplateContext>;
   @Input() thumbsOrientation: OrientationFlag;
   @Input() aria: Aria;
   @Input()
@@ -157,7 +159,7 @@ export class ViewerComponent implements OnChanges, OnInit, OnDestroy {
 
     if (isBrowser && this.mouseGestures) {
       this._zone.runOutsideAngular(() => {
-        const imageList = this.itemListRef.nativeElement;
+        const hostEl = this._hostRef.nativeElement;
         let mousedown: MouseEvent;
         let maxDeltaX = 0;
         let maxDeltaY = 0;
@@ -197,20 +199,20 @@ export class ViewerComponent implements OnChanges, OnInit, OnDestroy {
 
         const ondragstart = (e: DragEvent) => e.preventDefault();
 
-        imageList.addEventListener('mousedown', onmousedown, listenerOpts);
-        imageList.addEventListener('click', onclick, { capture: true });
-        imageList.addEventListener('dragstart', ondragstart);
+        hostEl.addEventListener('mousedown', onmousedown, listenerOpts);
+        hostEl.addEventListener('click', onclick, { capture: true });
+        hostEl.addEventListener('dragstart', ondragstart);
         this._destroy$.subscribe(() => {
-          imageList.removeEventListener('mousedown', onmousedown);
-          imageList.removeEventListener('click', onclick);
-          imageList.removeEventListener('dragstart', ondragstart);
+          hostEl.removeEventListener('mousedown', onmousedown);
+          hostEl.removeEventListener('click', onclick);
+          hostEl.removeEventListener('dragstart', ondragstart);
         });
       });
     }
 
     if (isBrowser && this.touchGestures) {
       this._zone.runOutsideAngular(() => {
-        const imageList = this.itemListRef.nativeElement;
+        const hostEl = this._hostRef.nativeElement;
         let horizontal = null;
         let touchstart: TouchEvent;
         let lastTouchmove: TouchEvent;
@@ -263,13 +265,13 @@ export class ViewerComponent implements OnChanges, OnInit, OnDestroy {
           lastTouchmove = null;
         };
 
-        imageList.addEventListener('touchstart', ontouchstart, listenerOpts);
+        hostEl.addEventListener('touchstart', ontouchstart, listenerOpts);
         document.addEventListener('touchmove', ontouchmove, {
           passive: !UA.ios,
         });
         document.addEventListener('touchend', ontouchend, listenerOpts);
         this._destroy$.subscribe(() => {
-          imageList.removeEventListener('touchstart', ontouchstart);
+          hostEl.removeEventListener('touchstart', ontouchstart);
           document.removeEventListener('touchmove', ontouchmove);
           document.removeEventListener('touchend', ontouchend);
         });